feat: Add traffic congestion metric and UI display
Introduces the `trafficCongestion` metric to the city stats. This metric will be used to represent the level of traffic within the city.

The UI overlay has been updated to display the traffic congestion alongside pollution, using a car icon. The `types.ts` file has been updated to include `trafficCongestion` in the `CityStats` interface.